src/requirements.txt
Django==2.0.6
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
mysqlclient==1.3.12
Pillow==5.1.0
PyJWT==1.6.4
pytz==2018.4
django-rest-swagger==2.2.0
sorl-thumbnail-serializer-field==0.2.1
django-filter==1.1.0
